requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# WebSocket Support
websockets==10.3
websocket-client==1.9.0

# Deriv API
git+https://github.com/deriv-com/python-deriv-api.git@7e8dd0f0920d4eac0fb904891e4005178a30ee19

# Data Processing
pandas==2.3.3
numpy==2.2.6

# Technical Indicators
ta==0.11.0
pandas-ta==0.4.71b0

# Telegram
python-telegram-bot==22.5

# Environment Variables
python-dotenv==1.2.1

# HTTP Client
httpx==0.28.1

# Pydantic for validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Async support
aiofiles==25.1.0

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
email-validator==2.3.0